Move state to game
